+++ lib/dradis/plugins/csv_export/gem_version.rb
@@ -1,6 +1,6 @@
 module Dradis
   module Plugins
-    module CSV
+    module CSVExport
       # Returns the version of the currently loaded CSV as a <tt>Gem::Version</tt>
       def self.gem_version
         Gem::Version.new VERSION::STRING